Reorganize files and folders.
