+++ include/gendil/Algebra/SparseMatrixTypes/COO/coomatrixapply.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "gendil/Algebra/SparseMatrixTypes/cscmatrixstorage.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/COO/coomatrixstorage.hpp"
 #include "gendil/Algebra/SparseMatrixTypes/sparseapplyarithmetic.hpp"
 #include "gendil/Algebra/vector.hpp"
 #include "gendil/Utilities/MathHelperFunctions/atomicadd.hpp"
@@ -18,17 +18,17 @@ namespace details
 {
 
 template < typename Matrix >
-void CheckCSCApplyDimensions(
+void CheckCOOApplyDimensions(
    const Matrix & matrix,
    const size_t x_size,
    const size_t y_size )
 {
    GENDIL_VERIFY(
       x_size == static_cast< size_t >( matrix.num_cols ),
-      "Apply(CSC backend, ...) input vector has the wrong size." );
+      "Apply(COO backend, ...) input vector has the wrong size." );
    GENDIL_VERIFY(
       y_size == static_cast< size_t >( matrix.num_rows ),
-      "Apply(CSC backend, ...) output vector has the wrong size." );
+      "Apply(COO backend, ...) output vector has the wrong size." );
 }
 
 template <
@@ -36,7 +36,7 @@ template <
    typename Matrix,
    typename InputValue,
    typename OutputValue >
-void ApplyHostCSCToRawPointers(
+void ApplyHostCOOToRawPointers(
    const Backend &,
    const Matrix & matrix,
    const InputValue * x_data,
@@ -55,32 +55,29 @@ void ApplyHostCSCToRawPointers(
       OutputValueType,
       ComputeType >();
 
+   #pragma omp parallel for
    for ( IndexType row = 0; row < matrix.num_rows; ++row )
    {
       y_data[row] = OutputValue( 0 );
    }
 
-   for ( IndexType col = 0; col < matrix.num_cols; ++col )
+   #pragma omp parallel for
+   for ( IndexType entry = 0; entry < matrix.nnz; ++entry )
    {
-      const ComputeType x_col =
+      const IndexType row = matrix.rows[entry];
+      const IndexType col = matrix.cols[entry];
+      const ComputeType contribution =
+         static_cast< ComputeType >( matrix.values[entry] ) *
          static_cast< ComputeType >( x_data[col] );
-
-      for ( IndexType entry = matrix.col_ptr[col];
-            entry < matrix.col_ptr[col + 1];
-            ++entry )
-      {
-         const IndexType row = matrix.row_ind[entry];
-         const ComputeType contribution =
-            static_cast< ComputeType >( matrix.values[entry] ) * x_col;
-         y_data[row] += static_cast< OutputValueType >( contribution );
-      }
+      #pragma omp atomic
+      y_data[row] += static_cast< OutputValueType >( contribution );
    }
 }
 
 #if defined(GENDIL_USE_DEVICE)
 template < typename ValueType, typename IndexType >
 __global__
-void CSCDeviceZeroKernel(
+void COODeviceZeroKernel(
    ValueType * y,
    const IndexType num_rows )
 {
@@ -105,7 +102,7 @@ template <
    typename InputValue,
    typename OutputValue >
 __global__
-void CSCDeviceApplyKernel(
+void COODeviceApplyKernel(
    const Backend backend,
    const Matrix matrix,
    const InputValue * x,
@@ -128,30 +125,25 @@ void CSCDeviceApplyKernel(
       static_cast< IndexType >( blockDim.x ) *
       static_cast< IndexType >( gridDim.x );
 
-   for ( IndexType col =
+   for ( IndexType entry =
             static_cast< IndexType >( blockIdx.x ) *
             static_cast< IndexType >( blockDim.x ) +
             static_cast< IndexType >( threadIdx.x );
-         col < matrix.num_cols;
-         col += stride )
+         entry < matrix.nnz;
+         entry += stride )
    {
-      const ComputeType x_col =
+      const IndexType row = matrix.rows[entry];
+      const IndexType col = matrix.cols[entry];
+      const ComputeType contribution =
+         static_cast< ComputeType >( matrix.values[entry] ) *
          static_cast< ComputeType >( x[col] );
-
-      for ( IndexType entry = matrix.col_ptr[col];
-            entry < matrix.col_ptr[col + 1];
-            ++entry )
-      {
-         const ComputeType contribution =
-            static_cast< ComputeType >( matrix.values[entry] ) * x_col;
-         AtomicAdd(
-            y[matrix.row_ind[entry]],
-            static_cast< OutputValueType >( contribution ) );
-      }
+      AtomicAdd(
+         y[row],
+         static_cast< OutputValueType >( contribution ) );
    }
 }
 
-inline dim3 MakeCSCApplyGrid(
+inline dim3 MakeCOOApplyGrid(
    const GlobalIndex work_items,
    const char * error_message )
 {
@@ -172,7 +164,7 @@ template <
    typename Matrix,
    typename InputValue,
    typename OutputValue >
-void ApplyDeviceCSCToRawPointers(
+void ApplyDeviceCOOToRawPointers(
    const Backend & backend,
    const Matrix & matrix,
    const InputValue * x_data,
@@ -184,35 +176,35 @@ void ApplyDeviceCSCToRawPointers(
    if ( matrix.num_rows > 0 )
    {
       const dim3 grid_dim =
-         MakeCSCApplyGrid(
+         MakeCOOApplyGrid(
             static_cast< GlobalIndex >( matrix.num_rows ),
-            "Apply(NativeDeviceCSCBackend, ...) zero launch grid is too large." );
+            "Apply(NativeDeviceCOOBackend, ...) zero launch grid is too large." );
       CheckDeviceLaunchConfiguration( grid_dim, block_dim, 0 );
       GENDIL_CHECK_NO_PENDING_DEVICE_ERROR(
-         "Apply(NativeDeviceCSCBackend, ...): before zero launch" );
-      CSCDeviceZeroKernel<<< grid_dim, block_dim >>>(
+         "Apply(NativeDeviceCOOBackend, ...): before zero launch" );
+      COODeviceZeroKernel<<< grid_dim, block_dim >>>(
          y_data,
          matrix.num_rows );
       GENDIL_CHECK_LAST_DEVICE_LAUNCH(
-         "Apply(NativeDeviceCSCBackend, ...) zero" );
+         "Apply(NativeDeviceCOOBackend, ...) zero" );
    }
 
-   if ( matrix.num_cols > 0 )
+   if ( matrix.nnz > 0 )
    {
       const dim3 grid_dim =
-         MakeCSCApplyGrid(
-            static_cast< GlobalIndex >( matrix.num_cols ),
-            "Apply(NativeDeviceCSCBackend, ...) apply launch grid is too large." );
+         MakeCOOApplyGrid(
+            static_cast< GlobalIndex >( matrix.nnz ),
+            "Apply(NativeDeviceCOOBackend, ...) apply launch grid is too large." );
       CheckDeviceLaunchConfiguration( grid_dim, block_dim, 0 );
       GENDIL_CHECK_NO_PENDING_DEVICE_ERROR(
-         "Apply(NativeDeviceCSCBackend, ...): before apply launch" );
-      CSCDeviceApplyKernel<<< grid_dim, block_dim >>>(
+         "Apply(NativeDeviceCOOBackend, ...): before apply launch" );
+      COODeviceApplyKernel<<< grid_dim, block_dim >>>(
          backend,
          matrix,
          x_data,
          y_data );
       GENDIL_CHECK_LAST_DEVICE_LAUNCH(
-         "Apply(NativeDeviceCSCBackend, ...) apply" );
+         "Apply(NativeDeviceCOOBackend, ...) apply" );
    }
 }
 #endif
@@ -225,12 +217,12 @@ template <
    typename MatrixBackend,
    typename BackendComputeType >
 void Apply(
-   const HostCSCBackend< BackendComputeType > & backend,
-   const CSCMatrix< ValueType, IndexType, MatrixBackend > & matrix,
+   const HostCOOBackend< BackendComputeType > & backend,
+   const COOMatrix< ValueType, IndexType, MatrixBackend > & matrix,
    const Vector & x,
    Vector & y )
 {
-   details::CheckCSCApplyDimensions(
+   details::CheckCOOApplyDimensions(
       matrix,
       x.Size(),
       y.Size() );
@@ -238,7 +230,7 @@ void Apply(
    const auto * x_data = x.ReadHostData();
    auto * y_data = y.WriteHostData();
 
-   details::ApplyHostCSCToRawPointers( backend, matrix, x_data, y_data );
+   details::ApplyHostCOOToRawPointers( backend, matrix, x_data, y_data );
 }
 
 template <
@@ -247,13 +239,13 @@ template <
    typename MatrixBackend,
    typename BackendComputeType >
 void Apply(
-   const NativeDeviceCSCBackend< BackendComputeType > & backend,
-   const CSCMatrix< ValueType, IndexType, MatrixBackend > & matrix,
+   const NativeDeviceCOOBackend< BackendComputeType > & backend,
+   const COOMatrix< ValueType, IndexType, MatrixBackend > & matrix,
    const Vector & x,
    Vector & y )
 {
 #if defined(GENDIL_USE_DEVICE)
-   details::CheckCSCApplyDimensions(
+   details::CheckCOOApplyDimensions(
       matrix,
       x.Size(),
       y.Size() );
@@ -261,12 +253,12 @@ void Apply(
    const auto * x_data = x.ReadDeviceData();
    auto * y_data = y.WriteDeviceData();
 
-   details::ApplyDeviceCSCToRawPointers( backend, matrix, x_data, y_data );
+   details::ApplyDeviceCOOToRawPointers( backend, matrix, x_data, y_data );
 #else
    static_assert(
-      dependent_false_v< CSCMatrix< ValueType, IndexType, MatrixBackend > >,
-      "Apply(NativeDeviceCSCBackend, ...) requires GENDIL_USE_DEVICE "
-      "(CUDA or HIP). Use CSCMatrix::operator() for CPU execution." );
+      dependent_false_v< COOMatrix< ValueType, IndexType, MatrixBackend > >,
+      "Apply(NativeDeviceCOOBackend, ...) requires GENDIL_USE_DEVICE "
+      "(CUDA or HIP). Use COOMatrix::operator() for CPU execution." );
 #endif
 }
 
@@ -277,12 +269,12 @@ template <
    typename MatrixBackend,
    typename BackendComputeType >
 void Apply(
-   const HostCSCBackend< BackendComputeType > & backend,
-   const CSCMatrix< ValueType, IndexType, MatrixBackend > & matrix,
+   const HostCOOBackend< BackendComputeType > & backend,
+   const COOMatrix< ValueType, IndexType, MatrixBackend > & matrix,
    const mfem::Vector & x,
    mfem::Vector & y )
 {
-   details::CheckCSCApplyDimensions(
+   details::CheckCOOApplyDimensions(
       matrix,
       static_cast< size_t >( x.Size() ),
       static_cast< size_t >( y.Size() ) );
@@ -290,7 +282,7 @@ void Apply(
    const auto * x_data = x.HostRead();
    auto * y_data = y.HostWrite();
 
-   details::ApplyHostCSCToRawPointers( backend, matrix, x_data, y_data );
+   details::ApplyHostCOOToRawPointers( backend, matrix, x_data, y_data );
 }
 
 template <
@@ -299,13 +291,13 @@ template <
    typename MatrixBackend,
    typename BackendComputeType >
 void Apply(
-   const NativeDeviceCSCBackend< BackendComputeType > & backend,
-   const CSCMatrix< ValueType, IndexType, MatrixBackend > & matrix,
+   const NativeDeviceCOOBackend< BackendComputeType > & backend,
+   const COOMatrix< ValueType, IndexType, MatrixBackend > & matrix,
    const mfem::Vector & x,
    mfem::Vector & y )
 {
 #if defined(GENDIL_USE_DEVICE)
-   details::CheckCSCApplyDimensions(
+   details::CheckCOOApplyDimensions(
       matrix,
       static_cast< size_t >( x.Size() ),
       static_cast< size_t >( y.Size() ) );
@@ -313,11 +305,11 @@ void Apply(
    const auto * x_data = x.Read();
    auto * y_data = y.Write();
 
-   details::ApplyDeviceCSCToRawPointers( backend, matrix, x_data, y_data );
+   details::ApplyDeviceCOOToRawPointers( backend, matrix, x_data, y_data );
 #else
    static_assert(
-      dependent_false_v< CSCMatrix< ValueType, IndexType, MatrixBackend > >,
-      "Apply(NativeDeviceCSCBackend, mfem::Vector, ...) requires "
+      dependent_false_v< COOMatrix< ValueType, IndexType, MatrixBackend > >,
+      "Apply(NativeDeviceCOOBackend, mfem::Vector, ...) requires "
       "GENDIL_USE_DEVICE (CUDA or HIP)." );
 #endif
 }
@@ -328,7 +320,7 @@ template <
    typename IndexType,
    typename MatrixBackend >
 template < typename InputVector, typename OutputVector >
-void CSCMatrix< ValueType, IndexType, MatrixBackend >::operator()(
+void COOMatrix< ValueType, IndexType, MatrixBackend >::operator()(
    const InputVector & x,
    OutputVector & y ) const
 {
@@ -344,18 +336,18 @@ template <
    typename OutputVector >
 void Apply(
    const Backend &,
-   const CSCMatrix< ValueType, IndexType, MatrixBackend > &,
+   const COOMatrix< ValueType, IndexType, MatrixBackend > &,
    const InputVector &,
    OutputVector & )
 {
    static_assert(
       dependent_false_v<
          Backend,
-         CSCMatrix< ValueType, IndexType, MatrixBackend >,
+         COOMatrix< ValueType, IndexType, MatrixBackend >,
          InputVector,
          OutputVector >,
-      "No CSCMatrix Apply overload is available for this backend/vector "
-      "combination. Include cscmatrixapply.hpp and use a supported CSC "
+      "No COOMatrix Apply overload is available for this backend/vector "
+      "combination. Include coomatrixapply.hpp and use a supported COO "
       "backend/vector type." );
 }
