+++ include/gendil/Algebra/SparseMatrixTypes/CSC/cscmatrix.hpp
@@ -5,5 +5,5 @@
 #pragma once
 
 // Public CSC matrix interface: storage plus Apply support.
-#include "gendil/Algebra/SparseMatrixTypes/cscmatrixstorage.hpp"
-#include "gendil/Algebra/SparseMatrixTypes/cscmatrixapply.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/CSC/cscmatrixstorage.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/CSC/cscmatrixapply.hpp"
