+++ include/gendil/Algebra/SparseMatrixTypes/HypreCSR/hyprecsrmatrix.hpp
@@ -7,7 +7,7 @@
 #ifdef GENDIL_USE_HYPRE
 
 // Public Hypre-compatible CSR matrix interface: storage plus native Apply support.
-#include "gendil/Algebra/SparseMatrixTypes/hyprecsrmatrixstorage.hpp"
-#include "gendil/Algebra/SparseMatrixTypes/hyprecsrmatrixapply.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/HypreCSR/hyprecsrmatrixstorage.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/HypreCSR/hyprecsrmatrixapply.hpp"
 
 #endif // GENDIL_USE_HYPRE
