+++ include/gendil/Algebra/SparseMatrixTypes/HypreCSR/hyprecsrmatrixapply.hpp
@@ -6,8 +6,8 @@
 
 #ifdef GENDIL_USE_HYPRE
 
-#include "gendil/Algebra/SparseMatrixTypes/csrmatrixapply.hpp"
-#include "gendil/Algebra/SparseMatrixTypes/hyprecsrmatrixstorage.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/CSR/csrmatrixapply.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/HypreCSR/hyprecsrmatrixstorage.hpp"
 #include "gendil/Interfaces/Hypre/hyprematvec.hpp"
 
 namespace gendil
