+++ include/gendil/FiniteElementMethod/MatrixAssembly/COO/coofinalization.hpp
@@ -5,9 +5,9 @@
 #pragma once
 
 #include "gendil/prelude.hpp"
-#include "gendil/Algebra/SparseMatrixTypes/coomatrix.hpp"
-#include "gendil/Algebra/SparseMatrixTypes/rawcootripletbuffer.hpp"
-#include "gendil/FiniteElementMethod/MatrixAssembly/rawcoosortreduce.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/COO/coomatrix.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/COO/rawcootripletbuffer.hpp"
+#include "gendil/FiniteElementMethod/MatrixAssembly/COO/rawcoosortreduce.hpp"
 
 namespace gendil {
