+++ include/gendil/FiniteElementMethod/MatrixAssembly/COO/rawcoosortreduce.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "gendil/prelude.hpp"
-#include "gendil/Algebra/SparseMatrixTypes/rawcootripletbuffer.hpp"
+#include "gendil/Algebra/SparseMatrixTypes/COO/rawcootripletbuffer.hpp"
 
 #include <algorithm>
 #include <vector>
