+++ include/gendil/FiniteElementMethod/MatrixAssembly/Generic/weakformtraversal.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "gendil/prelude.hpp"
-#include "gendil/FiniteElementMethod/MatrixAssembly/localdoforientation.hpp"
+#include "gendil/FiniteElementMethod/MatrixAssembly/Generic/localdoforientation.hpp"
 #include "gendil/FiniteElementMethod/MatrixFreeOperators/GenericOperator/genericoperator.hpp"
 #include "gendil/FiniteElementMethod/MatrixFreeOperators/KernelOperators/DoFIO/localdofoperations.hpp"
 #include "gendil/FiniteElementMethod/MatrixFreeOperators/KernelOperators/LoopHelpers/faceloop.hpp"
