--- include/gendil/FiniteElementMethod/MatrixFreeOperators/GenericOperator/operatorcontext.hpp
@@ -1,9 +1,0 @@
-// Copyright GenDiL Project Developers. See COPYRIGHT file for details.
-//
-// SPDX-License-Identifier: (BSD-3-Clause)
-
-#pragma once
-
-#include "gendil/FiniteElementMethod/MatrixFreeOperators/GenericOperator/operatorcontextcommon.hpp"
-#include "gendil/FiniteElementMethod/MatrixFreeOperators/GenericOperator/localoperatorcontext.hpp"
-#include "gendil/FiniteElementMethod/MatrixFreeOperators/GenericOperator/globaloperatorcontext.hpp"
